requirements.txt
asgiref==3.4.1
bleach==4.1.0
Django==3.2.6
django-environ==0.8.1
django-markdownify==0.9.0
django-markdownx==3.0.1
django-taggit==2.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.13.1
gunicorn==19.9.0
importlib-metadata==4.9.0
Markdown==3.3.6
packaging==21.3
Pillow==8.4.0
psycopg2==2.9.2
Pygments==2.10.0
pyparsing==3.0.6
python-slugify==5.0.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
webencodings==0.5.1
whitenoise==5.3.0
zipp==3.6.0
